Boot/Kernel: Fixup of argument passing from bootloader to kernel.

+++ src/kernel/kernel.cpp
@@ -5,7 +5,7 @@
 #include "system/common.hpp"
 #include "utility/log.hpp"
 
-extern "C" void kmain(u16 bootloaderLinesPrinted, MemoryMap* memoryMap)
+extern "C" void kmain(u32 bootloaderLinesPrinted, MemoryMap* memoryMap)
 {
 	move_cursor(bootloaderLinesPrinted + 1, 0);
 	log_info("PandaOS kernel!");
